integrated 0D and 1D CI functionality

+++ spm1d/data/uv0d/ci1.py
@@ -9,7 +9,7 @@ class MinnesotaGeyerRate(_base.DatasetCI1):
 	def _set_values(self):
 		self.www      = 'http://www.stat.umn.edu/geyer/3011/examp/conf.html'
 		self.datafile = 'http://www.stat.umn.edu/geyer/3011/mdata/chap16/eg16-01.dat'
-		self.Y        = np.array([29, 27, 34, 40, 22, 28, 14, 35, 26, 35, 12, 30, 23, 18, 11, 22, 23, 33])
+		self.Y        = np.array([29, 27, 34, 40, 22, 28, 14, 35, 26, 35, 12, 30, 23, 18, 11, 22, 23, 33], dtype=float)
 		self.alpha    = 0.05
 		self.mu       = None
 		self.ci       = (21.52709, 29.80625)
